fix: properly implement ResultingCode and add vuepress styling
