--- code-samples/additional-info/proxy/14/rest-proxy.js
@@ -1,10 +1,0 @@
-const { ProxyAgent } = require('undici');
-const { Client } = require('discord.js');
-
-// eslint-disable-next-line no-unused-vars
-const myClient = new Client({
-	// other client options
-	rest: {
-		agent: new ProxyAgent('http://my-proxy-server:port'),
-	},
-});
